Clerk sign in and out auth complete
